Crate ab_blake3

Optimized and more exotic APIs around BLAKE3

Constants

BLOCK_LEN

The number of bytes in a block

CHUNK_LEN

The number of bytes in a chunk, 1024.

KEY_LEN

The number of bytes in a key

OUT_LEN

The number of bytes in a hash

Functions

const_derive_key

const_hash

Hashing function like [blake3::hash()], but const fn

const_keyed_hash

The keyed hash function like [blake3::keyed_hash()], but const fn

le_bytes_from_words_32

Converts u32 words into bytes, the size matches BLAKE3 hash

single_block_derive_key

single_block_hash

Hashing function for at most single block worth of bytes.

single_block_hash_many_exact

Hashing function for many single-block inputs

single_block_hash_portable_words

Hashing function for at most single block worth of words using portable implementation.

single_block_keyed_hash

The keyed hash function for at most single block worth of bytes.

single_block_keyed_hash_many_exact

Keyed hash function for many single-block inputs

single_chunk_derive_key

single_chunk_hash

Hashing function for at most single chunk (CHUNK_LEN) worth of bytes.

single_chunk_keyed_hash

The keyed hash function for at most a single chunk (CHUNK_LEN) worth of bytes.

words_from_le_bytes_32

Converts bytes into u32 words, the size matches BLAKE3 hash

words_from_le_bytes_64

Converts bytes into u32 words, the size matches BLAKE3 block